API Project Test
I used a service based approach in the design of my application and built loosely coupled reusable helpers for performing common tasks.
All calls to articles go through a parent ArticleService which checks the application mode and in turn calls the appropriate service for fetching the article content.
Unfortunately, I did not get time to write the feature tests.